Run `on_docker_event` callbacks on main thread.
Use `asyncio.run_coroutine_threadsafe` to run
`on_docker_event.trigger` in the main event loop.

This way the docker events are monitored asynchronously on
the separate thread, but the callbacks are executed in the main thread.

This is necessary to avoid thread safety issues when fixture values are modified within the callback functions.