Add `__follow_wrapped__` flag.
Indicates that the signature of the wrapped function should be used when
binding a fixture.

If set to `True`, then the signature of the base function will be used.
If set to `False`, then the signature of the wrapper function will be
used.

If the flag is not present then it will be assumed `True`.