refactor: update push_status function
refactor: update state constants to match `virtool/virtool`

test: add test for status update behaviour

test: setup integration tests in pytest

In order to verify that status updates are in fact received by the
`jobs-api` server, it is necessary to access the database to view the
added status entries.

This commit sets up an integration testing environment in pytest,
allowing the database to be initialized before each test and it's state
to be queried afterwards.

test: connect to services using pytest-docker-compose

test: add fixtures for testing jobs

refactor: update push_status to use `step_name` and `step_description`

refactor: push status on workflow step start

refactor: utilize `on_workflow_start` hook to initialize the fixture scope

refactor: return the `WorkflowStep` object from the `step` decorator

test: use dev mode on virtool server

test: ensure `complete` status sent

refactor: log out the response text when an HTTP error code is received