test: setup integration testing in pytest
test: rename `jobs_api_url` fixture to `jobs_api`

The fixture name `jobs_api_url` conflicts with the `jobs_api_url`
fixture used to provide access to the fake API during unit tests.

test: remove `does_push_step_updates`

This functionality is covered by the integration tests for status
updates.

test: include all pytest plugins in root conftest.py

This is a requirement of pytest.

test: move `integration` to `tests/integration`

test: wait for redis connection to close before exiting from fixture

test: explicitly clear hooks before executing tests

test: add compose file to run tests in container

Accomplished by mounting the docker socket and running the test
container with `network-mode: "host"`.

This allows the `pytest-docker-compose` plugin to make the required
requests to the docker API, while running in the `workflow-tools`
container.

test: update github action

test: let pytest-docker-compose start services

test: remove old integration tests from ci

Integration tests can now be ran alongside the unit tests. The previous
integration tests will be re-factored to run in pytest instead.